The merkle signature scheme commits on a raw  representation of the public keys.
