+++ crypto/sig_abstractions.go
@@ -69,6 +69,8 @@ type Verifier interface {
 	// Verify and VerifyBytes returns error on bad signature, and any other problem.
 	Verify(message Hashable, sig ByteSignature) error
 	VerifyBytes(message []byte, sig ByteSignature) error
+	// GetRawVerificationBytes is used to fetch a plain serialized version of the public data (without the use of the message pack).
+	GetRawVerificationBytes() []byte
 }
 
 // GenericSigningKey holds a Signer, and the type of algorithm the Signer conforms with.
